+++ packages/node/src/app.module.ts
@@ -7,7 +7,6 @@ import { ScheduleModule } from '@nestjs/schedule';
 import { DbModule } from '@subql/node-core';
 import { ConfigureModule } from './configure/configure.module';
 import { FetchModule } from './indexer/fetch.module';
-import { RuntimeService } from './indexer/runtimeService';
 import { MetaModule } from './meta/meta.module';
 
 @Module({
